Fix FocusForge compile and startup compatibility

+++ FocusForge/src/main/java/bd/edu/seu/classproject/focusforge/exams/ExamScheduleService.java
@@ -1,15 +1,17 @@
 package bd.edu.seu.classproject.focusforge.exams;
 
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
-@RequiredArgsConstructor
 public class ExamScheduleService {
 
     private final ExamScheduleRepository examScheduleRepository;
+
+    public ExamScheduleService(ExamScheduleRepository examScheduleRepository) {
+        this.examScheduleRepository = examScheduleRepository;
+    }
 
     public ExamSchedule saveExam(ExamSchedule examSchedule) {
         return examScheduleRepository.save(examSchedule);
@@ -20,6 +22,8 @@ public class ExamScheduleService {
     }
 
     public void deleteById(Long id) {
-        examScheduleRepository.deleteById(id);
+        if (id != null && examScheduleRepository.existsById(id)) {
+            examScheduleRepository.deleteById(id);
+        }
     }
 }
